# src/banking-transactions-openapi.yaml
openapi: 3.0.3
info:
  title: Banking Transactions API
  version: 1.1.0
  description: >-
    API for processing banking transactions including deposits, withdrawals,
    refunds, checks, and payments. Supports creation, listing, and retrieval,
    plus replacement and deletion of pending transactions.
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://sandbox.api.example.com/v1
    description: Sandbox server
security:
  - bearerAuth: []
paths:
...
  /accounts/{accountId}/transactions/deposits:
    get:
      tags: [Transactions]
      summary: List deposits
      operationId: listDeposits
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Transactions ordered by createdAt descending, then transactionId ascending. An empty collection returns an empty items array.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ProcessingError'
    post:
      tags:
        - Transactions
      summary: Create a deposit transaction
      description: Creates a deposit for the specified account. Requires a valid bearer JWT with the admin role. The amount must be greater than 0.01 and no more than 1,000,000.00.
      operationId: createDeposit
      parameters:
        - name: accountId
          in: path
          required: true
          description: Unique account identifier. Must be exactly 10 characters.
          schema:
            type: string
            minLength: 10
            maxLength: 10
            pattern: '^[A-Za-z0-9]{10}$'
      requestBody:
        required: true
        description: Deposit details. Amount and currency are required; source and description are optional.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
            example:
              amount: 125.50
              currency: USD
              description: Payroll deposit
              source: Employer
      responses:
        '201':
          description: Deposit created successfully.
          headers:
            Location:
              description: URI of the newly created deposit.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
              example:
                transactionId: deposits-123
                status: PENDING
                amount: 125.50
                currency: USD
                createdAt: '2026-06-01T12:00:00Z'
                description: Payroll deposit
                source: Employer
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: The bearer token is valid, but the authenticated user does not have the admin role. No response body is returned.
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/ProcessingError'
...
components:
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      description: Unique account identifier. Must be exactly 10 alphanumeric characters.
      schema:
        type: string
        minLength: 10
        maxLength: 10
        pattern: '^[A-Za-z0-9]{10}$'
...
    Limit:
      name: limit
      in: query
      description: Maximum number of transactions to return.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of transactions to skip.
      schema:
        type: integer
        minimum: 0
        default: 0
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PositiveAmount:
      type: number
      description: Amount must be greater than zero.
      format: decimal
      minimum: 0.01
      exclusiveMinimum: true
      multipleOf: 0.01
    DepositAmount:
      allOf:
        - $ref: '#/components/schemas/PositiveAmount'
      maximum: 1000000
      description: Deposit amount must be greater than 0 and no more than 1,000,000.00.
    TransactionReference:
      type: string
      description: Unique transaction reference.
      minLength: 1
      maxLength: 64
    TransactionRequestBase:
      type: object
      required:
        - amount
      properties:
        amount:
          $ref: '#/components/schemas/PositiveAmount'
        currency:
          type: string
          description: ISO 4217 currency code.
          example: USD
          pattern: '^[A-Z]{3}$'
        description:
          type: string
          maxLength: 200
          example: Monthly rent payment
    DepositRequest:
      allOf:
        - $ref: '#/components/schemas/TransactionRequestBase'
      type: object
      required:
        - amount
      properties:
        amount:
          $ref: '#/components/schemas/DepositAmount'
        source:
          type: string
          maxLength: 64
          description: Source of funds.
          example: External transfer
...
    TransactionListResponse:
      type: object
      required: [items, limit, offset, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Effective page size.
        offset:
          type: integer
          minimum: 0
          description: Effective number of transactions skipped.
        total:
          type: integer
          minimum: 0
          description: Total number of transactions for this account and type, excluding deleted transactions.
    TransactionResponse:
      type: object
      description: Transaction details, including supplied editable fields applicable to its transaction type.
      required:
        - transactionId
        - status
        - amount
        - currency
      properties:
        transactionId:
          $ref: '#/components/schemas/TransactionReference'
        status:
          type: string
          enum: [PENDING, COMPLETED, REJECTED, FAILED]
        amount:
          type: number
          format: decimal
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        description:
          type: string
          maxLength: 200
        source:
          type: string
          maxLength: 64
          description: Source of funds for deposits.
        channel:
          type: string
          enum: [ATM, BRANCH, MOBILE, ONLINE]
          description: Channel for withdrawals.
        originalTransactionId:
          $ref: '#/components/schemas/TransactionReference'
        reason:
          type: string
          maxLength: 200
          description: Reason for a refund.
        checkNumber:
          type: string
          minLength: 1
          maxLength: 20
        payee:
          type: string
          maxLength: 120
        payeeAccountId:
          type: string
          minLength: 10
          maxLength: 10
          pattern: '^[A-Za-z0-9]{10}$'
        reference:
          type: string
          maxLength: 64
        createdAt:
          type: string
          format: date-time
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      required:
        - code
        - message
        - type
        - requestId
        - timestamp
      properties:
        code:
          type: string
          description: Machine-readable error code.
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable summary of the problem.
          example: Request validation failed.
        type:
          type: string
          description: Error category.
          enum: [validation, processing, authorization, not_found, conflict]
          example: validation
        requestId:
          type: string
          description: Correlation ID for tracing the request.
          example: req_8fbd9dca
        timestamp:
          type: string
          format: date-time
          description: UTC timestamp when the error occurred.
        details:
          type: array
          description: Detailed validation or business rule errors.
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - field
        - issue
      properties:
        field:
          type: string
          description: Field or parameter name associated with the error.
          example: amount
        issue:
          type: string
          description: Specific validation or processing issue.
          example: must be greater than 0
        rejectedValue:
          type: string
          nullable: true
          description: Value that was provided and rejected.
          example: '0'
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed.
              type: validation
              requestId: req_8fbd9dca
              timestamp: '2026-09-15T12:00:00Z'
              details:
                - field: amount
                  issue: must be greater than 0
                  rejectedValue: '0'
                - field: accountId
                  issue: must be exactly 10 characters
                  rejectedValue: '12345'
    ProcessingError:
      description: Transaction processing failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: PROCESSING_ERROR
              message: Transaction could not be processed.
              type: processing
              requestId: req_8fbd9dca
              timestamp: '2026-09-15T12:00:00Z'
              details:
                - field: transaction
                  issue: downstream processor returned an error
                  rejectedValue: null
    UnauthorizedError:
      description: Authentication is required or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Missing or invalid bearer token.
              type: authorization
              requestId: req_8fbd9dca
              timestamp: '2026-09-15T12:00:00Z'
    NotFoundError:
      description: The requested account or resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: ACCOUNT_NOT_FOUND
              message: Account not found.
              type: not_found
              requestId: req_8fbd9dca
              timestamp: '2026-09-15T12:00:00Z'
    ConflictError:
      description: The operation conflicts with an existing state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: TRANSACTION_CONFLICT
              message: A duplicate or conflicting transaction already exists.
              type: conflict
              requestId: req_8fbd9dca
              timestamp: '2026-09-15T12:00:00Z'
...